Refactoring pt 1
