+++ api/handlers.go
@@ -508,12 +508,12 @@ func (si *ServerImplementation) fetchAssets(ctx context.Context, options idb.Ass
 			Deleted:          row.Deleted,
 			Params: generated.AssetParams{
 				Creator:       creator.String(),
-				Name:          strPtr(row.Params.AssetParams.AssetName),
-				NameB64:       bytePtr(row.Params.AssetNameBytes),
-				UnitName:      strPtr(row.Params.AssetParams.UnitName),
-				UnitNameB64:   bytePtr(row.Params.UnitNameBytes),
-				Url:           strPtr(row.Params.AssetParams.URL),
-				UrlB64:        bytePtr(row.Params.URLBytes),
+				Name:          strPtr(row.Params.AssetName),
+				NameB64:       bytePtr([]byte(row.Params.AssetName)),
+				UnitName:      strPtr(row.Params.UnitName),
+				UnitNameB64:   bytePtr([]byte(row.Params.UnitName)),
+				Url:           strPtr(row.Params.URL),
+				UrlB64:        bytePtr([]byte(row.Params.URL)),
 				Total:         row.Params.Total,
 				Decimals:      uint64(row.Params.Decimals),
 				DefaultFrozen: boolPtr(row.Params.DefaultFrozen),
