+++ idb/postgres/postgres_integration_test.go
@@ -1103,3 +1103,200 @@ func TestOpenDbAgain(t *testing.T) {
 	_, err = OpenPostgres(connStr, idb.IndexerDbOptions{}, nil)
 	require.NoError(t, err)
 }
+
+// TestEmbeddedNullString make sure we're able to import cheeky assets.
+func TestNonDisplayableUTF8(t *testing.T) {
+	tests := []struct {
+		Name              string
+		AssetName         string
+		AssetUnit         string
+		AssetURL          string
+		ExpectedAssetName string
+		ExpectedAssetUnit string
+		ExpectedAssetURL  string
+	}{
+		{
+			Name:              "Normal",
+			AssetName:         "asset-name",
+			AssetUnit:         "au",
+			AssetURL:          "https://algorand.com",
+			ExpectedAssetName: "asset-name",
+			ExpectedAssetUnit: "au",
+			ExpectedAssetURL:  "https://algorand.com",
+		},
+		{
+			Name:              "Embedded Null",
+			AssetName:         "asset\000name",
+			AssetUnit:         "a\000u",
+			AssetURL:          "https:\000//algorand.com",
+			ExpectedAssetName: "",
+			ExpectedAssetUnit: "",
+			ExpectedAssetURL:  "",
+		},
+		{
+			Name:              "Invalid UTF8",
+			AssetName:         "asset\x8cname",
+			AssetUnit:         "a\x8cu",
+			AssetURL:          "https:\x8c//algorand.com",
+			ExpectedAssetName: "",
+			ExpectedAssetUnit: "",
+			ExpectedAssetURL:  "",
+		},
+		{
+			Name:              "Emoji",
+			AssetName:         "💩",
+			AssetUnit:         "💰",
+			AssetURL:          "🌐",
+			ExpectedAssetName: "💩",
+			ExpectedAssetUnit: "💰",
+			ExpectedAssetURL:  "🌐",
+		},
+	}
+
+	assetID := uint64(1)
+	round := test.Round
+	var creator types.Address
+
+	for _, testcase := range tests {
+		testcase := testcase
+		name := testcase.AssetName
+		unit := testcase.AssetUnit
+		url := testcase.AssetURL
+		creator[0] = byte(assetID)
+
+		t.Run(testcase.Name, func(t *testing.T) {
+			t.Parallel()
+			db, shutdownFunc := setupIdb(t, test.MakeGenesis())
+			defer shutdownFunc()
+
+			txn, txnRow := test.MakeAssetConfigOrPanic(
+				round, 0, assetID, math.MaxUint64, 0, false, unit, name, url, test.AccountA)
+
+			// Test 1: import/accounting should work.
+			importTxns(t, db, round, txn)
+			accountTxns(t, db, round, txnRow)
+
+			// Test 2: asset results properly serialized
+			assets, _ := db.Assets(context.Background(), idb.AssetsQuery{AssetID: assetID})
+			num := 0
+			for asset := range assets {
+				require.NoError(t, asset.Error)
+				require.Equal(t, name, asset.Params.AssetName)
+				require.Equal(t, unit, asset.Params.UnitName)
+				require.Equal(t, url, asset.Params.URL)
+				num++
+			}
+			require.Equal(t, 1, num)
+
+			// Test 3: transaction results properly serialized
+			transactions, _ := db.Transactions(context.Background(), idb.TransactionFilter{})
+			num = 0
+			for tx := range transactions {
+				require.NoError(t, tx.Error)
+				// Note: These are created from the TxnBytes, so they have the exact name with embedded null.
+				var txn sdk_types.SignedTxn
+				require.NoError(t, msgpack.Decode(tx.TxnBytes, &txn))
+				require.Equal(t, name, txn.Txn.AssetParams.AssetName)
+				require.Equal(t, unit, txn.Txn.AssetParams.UnitName)
+				require.Equal(t, url, txn.Txn.AssetParams.URL)
+				num++
+			}
+			require.Equal(t, 1, num)
+
+			requireNilOrEqual := func(t *testing.T, expected string, actual *string) {
+				if expected == "" {
+					require.Nil(t, actual)
+				} else {
+					require.NotNil(t, actual)
+					require.Equal(t, expected, *actual)
+				}
+			}
+			// Test 4: account results should have the correct asset
+			accounts, _ := db.GetAccounts(context.Background(), idb.AccountQueryOptions{EqualToAddress: test.AccountA[:], IncludeAssetParams: true})
+			num = 0
+			for acct := range accounts {
+				require.NoError(t, acct.Error)
+				require.NotNil(t, acct.Account.CreatedAssets)
+				require.Len(t, *acct.Account.CreatedAssets, 1)
+
+				asset := (*acct.Account.CreatedAssets)[0]
+				if testcase.ExpectedAssetName == "" {
+					require.Nil(t, asset.Params.Name)
+				}
+				requireNilOrEqual(t, testcase.ExpectedAssetName, asset.Params.Name)
+				requireNilOrEqual(t, testcase.ExpectedAssetUnit, asset.Params.UnitName)
+				requireNilOrEqual(t, testcase.ExpectedAssetURL, asset.Params.Url)
+				require.Equal(t, []byte(name), *asset.Params.NameB64)
+				require.Equal(t, []byte(unit), *asset.Params.UnitNameB64)
+				require.Equal(t, []byte(url), *asset.Params.UrlB64)
+				num++
+			}
+			require.Equal(t, 1, num)
+		})
+	}
+}
+
+// TestReconfigAsset
+func TestReconfigAsset(t *testing.T) {
+	db, shutdownFunc := setupIdb(t, test.MakeGenesis())
+	defer shutdownFunc()
+
+	unit := "co\000in"
+	name := "algo"
+	url := "https://algorand.com"
+
+	assetID := uint64(1)
+	txn, txnRow := test.MakeAssetConfigOrPanic(
+		test.Round, 0, assetID, math.MaxUint64, 0, false, unit, name, url, test.AccountA)
+
+	txn2 := sdk_types.SignedTxnWithAD{
+		SignedTxn: sdk_types.SignedTxn{
+			Txn: sdk_types.Transaction{
+				Type: "acfg",
+				Header: sdk_types.Header{
+					Sender:     test.AccountA,
+					Fee:        sdk_types.MicroAlgos(1000),
+					FirstValid: sdk_types.Round(test.Round+1),
+					LastValid:  sdk_types.Round(test.Round+1),
+				},
+				AssetConfigTxnFields: sdk_types.AssetConfigTxnFields{
+					ConfigAsset: sdk_types.AssetIndex(assetID),
+					AssetParams: sdk_types.AssetParams{
+						Freeze:        test.AccountB,
+						Clawback:      test.AccountC,
+					},
+				},
+			},
+		},
+	}
+
+	txnRow2 := idb.TxnRow{
+		Round:    uint64(txn2.Txn.FirstValid),
+		TxnBytes: msgpack.Encode(txn2),
+		AssetID:  assetID,
+	}
+
+	importTxns(t, db, test.Round, txn)
+	accountTxns(t, db, test.Round, txnRow)
+	importTxns(t, db, test.Round+1, &txn2)
+	accountTxns(t, db, test.Round+1, &txnRow2)
+
+	// Test 2: asset results properly serialized
+	assets, _ := db.Assets(context.Background(), idb.AssetsQuery{AssetID: assetID})
+	num := 0
+	for asset := range assets {
+		require.NoError(t, asset.Error)
+		require.Equal(t, []byte(name), asset.Params.AssetName)
+		require.Equal(t, []byte(unit), asset.Params.UnitName)
+		require.Equal(t, []byte(url), asset.Params.URL)
+
+		// These were cleared out
+		require.Equal(t, sdk_types.ZeroAddress, asset.Params.Manager)
+		require.Equal(t, sdk_types.ZeroAddress, asset.Params.Reserve)
+		// These were updated
+		require.Equal(t, test.AccountB, asset.Params.Freeze)
+		require.Equal(t, test.AccountC, asset.Params.Clawback)
+		num++
+	}
+	require.Equal(t, 1, num)
+}
